requirements.txt
ultralytics==8.0.114
pandas==2.0.2
opencv-python==4.7.0.72
numpy==1.24.3
scipy==1.10.1
easyocr==1.7.0
filterpy==1.4.5
firebase-admin>=6.0.0
gradio>=3.50.0
streamlit>=1.28.0
python-dotenv>=1.0.0
pytest>=7.4.0


